# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Condominium Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/sync:
    post:
      operationId: syncUser
      tags: [auth]
      summary: Sync Clerk user to local database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSyncInput"
      responses:
        "200":
          description: User synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current user with company and unit context
      responses:
        "200":
          description: Current user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserContext"
        "401":
          description: Unauthenticated

  # ── Companies ─────────────────────────────────────────────────────────────
...
  /companies/{companyId}/buildings:
    get:
      operationId: getCompanyBuildings
      tags: [buildings]
      summary: List all buildings for a company
      parameters:
        - $ref: "#/components/parameters/companyId"
      responses:
        "200":
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
    post:
      operationId: createBuilding
      tags: [buildings]
      summary: Create a building
      parameters:
        - $ref: "#/components/parameters/companyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingInput"
      responses:
        "201":
          description: Building created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
...
  /buildings/{buildingId}:
    get:
      operationId: getBuilding
      tags: [buildings]
      summary: Get building details
      parameters:
        - $ref: "#/components/parameters/buildingId"
      responses:
        "200":
          description: Building details with unit count summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingDetail"
        "404":
          description: Not found
    patch:
      operationId: updateBuilding
      tags: [buildings]
      summary: Update building details
      parameters:
        - $ref: "#/components/parameters/buildingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingUpdate"
      responses:
        "200":
          description: Building updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
  /buildings/{buildingId}/units:
    get:
      operationId: getBuildingUnits
      tags: [units]
      summary: List units for a building
      parameters:
        - $ref: "#/components/parameters/buildingId"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, all]
          description: Filter by status (default active)
      responses:
        "200":
          description: List of units with ownership info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnitSummary"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit (apartment) in a building
      parameters:
        - $ref: "#/components/parameters/buildingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  # ── Units ─────────────────────────────────────────────────────────────────
...
components:
  parameters:
    companyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
    buildingId:
      name: buildingId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    UserSyncInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        fullName:
          type: ["string", "null"]
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        clerkUserId:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserContext:
      type: object
      required: [user, adminCompanies, ownedUnits, tenancy]
      properties:
        user:
          $ref: "#/components/schemas/User"
        adminCompanies:
          type: array
          items:
            $ref: "#/components/schemas/Company"
          description: Companies where the user is an administrator
        ownedUnits:
          type: array
          items:
            $ref: "#/components/schemas/OwnedUnitSummary"
          description: Units where the user is an owner
        tenancy:
          $ref: "#/components/schemas/TenancySummary"
          description: Active tenancy, or null

    # ── Company ──────────────────────────────────────────────────────────────
...
    Company:
      type: object
      required: [id, name, subscriptionTier, enterpriseFlagged, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: ["string", "null"]
        subscriptionTier:
          type: string
          enum: [free, standard, enterprise]
        enterpriseFlagged:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    BuildingInput:
      type: object
      required: [name, locality]
      properties:
        name:
          type: string
          minLength: 1
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        locality:
          type: string
        postcode:
          type: ["string", "null"]
        country:
          type: string
          default: MT
        status:
          type: string
          enum: [active, inactive]
          default: active
    BuildingUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        locality:
          type: string
        postcode:
          type: ["string", "null"]
        country:
          type: string
        status:
          type: string
          enum: [active, inactive]
    Building:
      type: object
      required: [id, companyId, name, locality, status, createdAt]
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        locality:
          type: string
        postcode:
          type: ["string", "null"]
        country:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BuildingDetail:
      allOf:
        - $ref: "#/components/schemas/Building"
        - type: object
          required: [activeUnitCount, archivedUnitCount]
          properties:
            activeUnitCount:
              type: integer
            archivedUnitCount:
              type: integer

    # ── Unit ─────────────────────────────────────────────────────────────────
    UnitInput:
      type: object
      required: [unitNumber]
      properties:
        unitNumber:
          type: string
          minLength: 1
          description: e.g. "Apt 1", "Unit 4A"
        unitType:
          type: string
          enum: [apartment, garage, commercial, other]
          default: apartment
        floor:
          type: ["integer", "null"]
...
    Unit:
      type: object
      required: [id, companyId, buildingId, unitNumber, unitType, status, createdAt]
      properties:
        id:
          type: string
        companyId:
          type: string
        buildingId:
          type: string
        unitNumber:
          type: string
        unitType:
          type: string
          enum: [apartment, garage, commercial, other]
        floor:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, archived]
        activatedAt:
          type: ["string", "null"]
          format: date-time
        archivedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UnitSummary:
      allOf:
        - $ref: "#/components/schemas/Unit"
        - type: object
          properties:
            ownerName:
              type: ["string", "null"]
            ownerEmail:
              type: ["string", "null"]
            ownerStatus:
              type: ["string", "null"]
              enum: [active, pending, null]
            tenantName:
              type: ["string", "null"]
            tenantStatus:
              type: ["string", "null"]
              enum: [active, pending, null]
...
    OwnedUnitSummary:
      type: object
      required: [membership, unit, building, company]
      properties:
        membership:
          $ref: "#/components/schemas/UnitMembership"
        unit:
          $ref: "#/components/schemas/Unit"
        building:
          $ref: "#/components/schemas/Building"
        company:
          $ref: "#/components/schemas/Company"
    TenancySummary:
      type: ["object", "null"]
      properties:
        membership:
          $ref: "#/components/schemas/UnitMembership"
        unit:
          $ref: "#/components/schemas/Unit"
        building:
          $ref: "#/components/schemas/Building"
        company:
          $ref: "#/components/schemas/Company"

    # ── Memberships & Invitations ─────────────────────────────────────────────
...
    UnitMembership:
      type: object
      required: [id, unitId, role, status, invitedName, invitedEmail, createdAt]
      properties:
        id:
          type: string
        unitId:
          type: string
        userId:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, tenant]
        status:
          type: string
          enum: [pending, active, revoked]
        invitedName:
          type: string
        invitedEmail:
          type: string
        invitationExpiresAt:
          type: ["string", "null"]
          format: date-time
        activatedAt:
          type: ["string", "null"]
          format: date-time
        revokedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...